gestionar pedido de tarea

+++ Mi_Ram_HQ/utils.h
@@ -22,6 +22,6 @@
 extern t_log* logger;
 extern t_config* config;
 
-size_t tamanio_tarea(t_tarea* tarea);
+int tamanio_tarea(t_tarea* tarea);
 
 #endif /* MI_RAM_HQ_UTILS_H_ */
